burn rate panel
