+++ x-pack/solutions/observability/plugins/slo/public/embeddable/slo/common/slo_overview_details.tsx
@@ -59,7 +59,12 @@ export function SloOverviewDetailsContent({
         ))}
       </EuiTabs>
       <EuiSpacer size="m" />
-      <SloDetails slo={slo} isAutoRefreshing={false} selectedTabId={selectedTabId} />
+      <SloDetails
+        slo={slo}
+        isAutoRefreshing={false}
+        selectedTabId={selectedTabId}
+        renderMode="flyout"
+      />
     </>
   );
 }
