+++ x-pack/solutions/observability/plugins/slo/public/pages/slo_details/components/burn_rate_panel/burn_rate_flyout_panel.tsx
@@ -1,0 +1,78 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { EuiButtonGroup, EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
+import { i18n } from '@kbn/i18n';
+import React from 'react';
+import { ErrorRateChart } from '../../../../components/slo/error_rate_chart';
+import { BurnRateStatus } from './burn_rate_status';
+import { burnRateWindowLabel } from './utils';
+import { SloFlyoutCard } from '../../shared_flyout/flyout_card';
+import { useBurnRatePanel } from './hooks/use_burn_rate_panel';
+import type { BurnRatePanelProps } from './types';
+
+export function BurnRateFlyoutPanel({ slo, isAutoRefreshing }: BurnRatePanelProps) {
+  const {
+    burnRateWindows,
+    selectedWindow,
+    onBurnRateOptionChange,
+    longWindowBurnRate,
+    shortWindowBurnRate,
+    isLoading,
+    threshold,
+    currentStatus,
+    dataTimeRange,
+  } = useBurnRatePanel({ slo, isAutoRefreshing });
+
+  return (
+    <SloFlyoutCard
+      title={i18n.translate('xpack.slo.burnRates.burnRatePanelTitle', {
+        defaultMessage: 'Burn rate',
+      })}
+      renderTooltip
+      data-test-subj="burnRatePanel"
+    >
+      <EuiFlexGroup direction="column" gutterSize="m">
+        <EuiFlexItem>
+          <EuiButtonGroup
+            legend={i18n.translate('xpack.slo.burnRate.timeRangeBtnLegend', {
+              defaultMessage: 'Select the time range',
+            })}
+            options={burnRateWindows.map((burnRateWindow) => ({
+              id: burnRateWindow.name,
+              label: burnRateWindowLabel(burnRateWindow),
+              'aria-label': burnRateWindowLabel(burnRateWindow),
+            }))}
+            idSelected={selectedWindow.name}
+            onChange={onBurnRateOptionChange}
+            isFullWidth
+            buttonSize="compressed"
+          />
+        </EuiFlexItem>
+        <EuiFlexGroup direction="row" gutterSize="m">
+          <EuiFlexItem grow={1}>
+            <BurnRateStatus
+              selectedWindow={selectedWindow}
+              shortWindowBurnRate={shortWindowBurnRate}
+              longWindowBurnRate={longWindowBurnRate}
+              isLoading={isLoading}
+            />
+          </EuiFlexItem>
+
+          <EuiFlexItem grow={3}>
+            <ErrorRateChart
+              slo={slo}
+              dataTimeRange={dataTimeRange}
+              threshold={threshold}
+              variant={currentStatus === 'BREACHED' ? 'danger' : 'success'}
+            />
+          </EuiFlexItem>
+        </EuiFlexGroup>
+      </EuiFlexGroup>
+    </SloFlyoutCard>
+  );
+}
